+++ frontend/src/layouts/DefaultLayout/index.tsx
@@ -1,6 +1,7 @@
-import MainPage from './MainPage';
 import { useEffect } from 'react';
 import { Outlet } from 'react-router-dom';
+
+import Landing from '@/pages/host/Landing';
 
 import useModal from '@/hooks/useModal';
 
@@ -14,7 +15,7 @@ const DefaultLayout: React.FC = () => {
     scrollTo(0, 0);
   }, []);
 
-  return <>{isRootPath ? <MainPage /> : <Outlet />}</>;
+  return <>{isRootPath ? <Landing /> : <Outlet />}</>;
 };
 
 export default DefaultLayout;
